Diff 1:
Changeset:

Implemented a functionality for tapping on a child UiElement by selector (used for tapping on buttons);
Added enum for the send notification button in the content descriptor class;
Implemented a functionality for starting the notification test activity;
Added the notification test activity in the android manifest;
Added a button for starting the notification test activity from the main activity;
Added a listener for tapping on the notification test activity button in the main activity.

Initial import:

Added an integration test for testing the interaction with the notification bar;
Added an XML for the layout of the notification test activity;
Added an activity for sending notification to the validator application for testing the interaction with the notification bar.

Note:

There are no integration tests for the pinchIn and pinchOut functions, since notification are UI elements and use the already implemented and tested functionalities for pinchOut and pinchIn.

[https://inttools.musala.com/reviewboard/r/28/ $28] see #3390

Diff 2:

Changeset:

Added description for parameters and returns in several functions in Bounds and Point;
Added more information in the description of the ServerConnectionProperties.

Diff 3:

Changeset:
* Added OpenQuickSettingsTest, OpenNotificationTest, NotificationInteractionTest to the integration tests.

git-svn-id: https://trac.musala.com/ARDC/repo@10295 34708651-a5da-c34d-b20c-10ddab38317d