Initial import:
Create a sample task, which finds the page of the Muffin Conference and likes it. Scrolling is not used in this task (it is already used when checking). Clear and inputing test is included in the scenario.

Changeset:
Add upper left corner of the edit text field as argument when double tapping.

[https://inttools.musala.com/reviewboard/r/185/ $185] see #3763

git-svn-id: https://trac.musala.com/ARDC/repo@10524 34708651-a5da-c34d-b20c-10ddab38317d