=== ChangeSet ===
Added method for getting device's screenshot and returning Image instance in order to ease the use of the Image class as getting the screenshot is the main source of Images.
Added method that validates if given image is on the active screen.

=== Initial Imports ===

Added unit test for the method that validates if given image is on the active screen.
Added image files in test-resources in order to test the method.

[https://inttools.musala.com/reviewboard/r/381/ $381] see #4129

git-svn-id: https://trac.musala.com/ARDC/repo@10899 34708651-a5da-c34d-b20c-10ddab38317d