Created UI Element Retrieving from the Device. Using the accessibility service.
Added method in UiAutomator for retrieving all ui elements on the screen matching the selector. Added integration tests for the new methods on the Screen and the Device. New activity in OnDeviceValidator containing visible and invisible elements was added in order to write an approproate integration test. Fixed getting the path of an element method that failed due to using != instead of equals.

See #9

Change-Id: Ia19fb595b1cbe618bb1cc9959ec1aa7a914f5fba